minute property type of presentation updated as integer

+++ src/main/java/io/github/demirmustafa/meetingapp/api/model/request/CreatePresentationRequest.java
@@ -17,7 +17,7 @@ public class CreatePresentationRequest {
     @NotNull
     private String name;
 
-    private Long minutes;
+    private Integer minutes;
 
     @NotNull
     private PresentationTimeType timeType;
